agent/requirements.txt
python-socketio[client]>=5.8.0
pyautogui>=0.9.54
requests>=2.28.0
